feat: Show new destinations on list page (#24961)
