+++ frontend/src/scenes/pipeline/destinations/NewDestinations.tsx
@@ -1,75 +1,35 @@
 import { IconPlusSmall } from '@posthog/icons'
-import { LemonBanner, LemonButton, LemonInput, LemonSelect, LemonTable, Link } from '@posthog/lemon-ui'
+import { LemonButton, LemonTable, Link } from '@posthog/lemon-ui'
 import { useActions, useValues } from 'kea'
-import { FlaggedFeature } from 'lib/components/FlaggedFeature'
 import { PayGateButton } from 'lib/components/PayGateMini/PayGateButton'
 import { PayGateMini } from 'lib/components/PayGateMini/PayGateMini'
-import { useFeatureFlag } from 'lib/hooks/useFeatureFlag'
 import { LemonTableLink } from 'lib/lemon-ui/LemonTable/LemonTableLink'
 
-import { sidePanelStateLogic } from '~/layout/navigation-3000/sidepanel/sidePanelStateLogic'
-import { AvailableFeature, PipelineStage, SidePanelTab } from '~/types'
+import { AvailableFeature, PipelineStage } from '~/types'
 
 import { pipelineAccessLogic } from '../pipelineAccessLogic'
-import { PipelineBackend } from '../types'
+import { DestinationsFilters } from './DestinationsFilters'
+import { destinationsFiltersLogic } from './destinationsFiltersLogic'
 import { DestinationTag } from './DestinationTag'
 import { newDestinationsLogic } from './newDestinationsLogic'
 
-export function DestinationOptionsTable(): JSX.Element {
-    const hogFunctionsEnabled = !!useFeatureFlag('HOG_FUNCTIONS')
-    const { loading, filteredDestinations, filters } = useValues(newDestinationsLogic)
-    const { setFilters, openFeedbackDialog } = useActions(newDestinationsLogic)
-    const { canEnableDestination } = useValues(pipelineAccessLogic)
-    const { openSidePanel } = useActions(sidePanelStateLogic)
-
+export function NewDestinations(): JSX.Element {
     return (
         <div className="space-y-2">
             <PayGateMini feature={AvailableFeature.DATA_PIPELINES} />
+            <DestinationsFilters hideShowPaused />
+            <DestinationOptionsTable />
+        </div>
+    )
+}
 
-            <FlaggedFeature flag="hog-functions" match={false}>
-                <LemonBanner
-                    type="info"
-                    action={{
-                        onClick: () => openSidePanel(SidePanelTab.FeaturePreviews),
-                        children: 'Enable feature preview',
-                    }}
-                >
-                    We're excited to announce <b>Destinations 3000</b> - the new version of our realtime destinations
-                    that include a range of pre-built templates, native filtering, templating and even customizing the
-                    code.
-                </LemonBanner>
-            </FlaggedFeature>
+export function DestinationOptionsTable(): JSX.Element {
+    const { loading, filteredDestinations, hiddenDestinations } = useValues(newDestinationsLogic)
+    const { canEnableDestination } = useValues(pipelineAccessLogic)
+    const { resetFilters } = useActions(destinationsFiltersLogic)
 
-            <div className="flex items-center gap-2">
-                <LemonInput
-                    type="search"
-                    placeholder="Search..."
-                    value={filters.search ?? ''}
-                    onChange={(e) => setFilters({ search: e })}
-                />
-                <Link className="text-sm font-semibold" subtle onClick={() => openFeedbackDialog()}>
-                    Can't find what you're looking for?
-                </Link>
-                <div className="flex-1" />
-                <LemonSelect
-                    type="secondary"
-                    size="small"
-                    options={
-                        [
-                            { label: 'All kinds', value: null },
-                            hogFunctionsEnabled
-                                ? { label: 'Realtime (new)', value: PipelineBackend.HogFunction }
-                                : undefined,
-                            hogFunctionsEnabled
-                                ? { label: 'Realtime (deprecated)', value: PipelineBackend.Plugin }
-                                : { label: 'Realtime', value: PipelineBackend.Plugin },
-                            { label: 'Batch exports', value: PipelineBackend.BatchExport },
-                        ].filter(Boolean) as { label: string; value: PipelineBackend | null }[]
-                    }
-                    value={filters.kind}
-                    onChange={(e) => setFilters({ kind: e ?? undefined })}
-                />
-            </div>
+    return (
+        <>
             <LemonTable
                 dataSource={filteredDestinations}
                 size="small"
@@ -103,7 +63,6 @@ export function DestinationOptionsTable(): JSX.Element {
                         },
                     },
                     {
-                        title: 'Actions',
                         width: 100,
                         align: 'right',
                         render: function RenderActions(_, target) {
@@ -127,6 +86,11 @@ export function DestinationOptionsTable(): JSX.Element {
                     },
                 ]}
             />
-        </div>
+            {hiddenDestinations.length > 0 && (
+                <div className="text-muted-alt">
+                    {hiddenDestinations.length} hidden. <Link onClick={() => resetFilters()}>Show all</Link>
+                </div>
+            )}
+        </>
     )
 }
